+++ src/index.js
@@ -1,1 +1,3 @@
-export { default } from './store';
+export { default as Store } from './store';
+export { default as combine } from './combine';
+export { default as utils } from './utils';