+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0 (2017-03-08)
+
+Features:
+- Minor version bump.
+- You should now `import { Store }` instead of `import Store`.
+- Add ability to combine reducers and effects with a given namespace.
+
 ## 1.0.0 (2017-03-04)
 
 Initial release.